refactor: repositioning server-cards
