+++ apps/dashboard/app/routes/dashboard/useDashboard.ts
@@ -1,4 +1,4 @@
-import { type Server } from "~/components/dashboard/ServerCardGrid";
+import { type Server } from "~/components/cards/server-card/ServerCardGrid";
 import { ServerStatusType } from "proto-gen-ts/daemon_pb";
 
 interface UseDashboard {
